@@ -5,29 +5,24 @@
   "dependencies": {
     "@emotion/react": "^11.9.0",
     "@emotion/styled": "^11.8.1",
-    "@google-pay/button-react": "^2.3.0",
-    "@material-ui/core": "^4.11.4",
-    "@material-ui/icons": "^4.11.2",
     "@mui/icons-material": "^5.6.2",
-    "@mui/material": "^5.6.3",
-    "@testing-library/jest-dom": "^5.12.0",
-    "@testing-library/react": "^11.2.6",
-    "@testing-library/user-event": "^12.8.3",
-    "axios": "^0.21.1",
-    "clsx": "^1.1.1",
-    "eslint": "^7.32.0",
-    "react": "^17.0.2",
+    "@mui/material": "^5.7.0",
+    "@mui/styles": "^5.7.0",
+    "@testing-library/jest-dom": "^5.16.4",
+    "@testing-library/react": "^13.2.0",
+    "@testing-library/user-event": "^13.5.0",
+    "axios": "^0.27.2",
+    "react": "18.1.0",
     "react-countdown": "^2.3.2",
-    "react-dom": "^17.0.2",
-    "react-material-ui-carousel": "^2.2.6",
-    "react-multi-carousel": "^2.6.2",
-    "react-redux": "^7.2.4",
-    "react-router-dom": "^5.2.0",
-    "react-scripts": "4.0.3",
-    "redux": "^4.1.0",
+    "react-dom": "^18.1.0",
+    "react-multi-carousel": "^2.8.0",
+    "react-redux": "^8.0.1",
+    "react-router-dom": "^6.3.0",
+    "react-scripts": "5.0.1",
+    "redux": "^4.2.0",
     "redux-devtools-extension": "^2.13.9",
-    "redux-thunk": "^2.3.0",
-    "web-vitals": "^1.1.2"
+    "redux-thunk": "^2.4.1",
+    "web-vitals": "^2.1.4"
   },
   "scripts": {
     "start": "react-scripts start",
