@@ -1,23 +1,22 @@
-import { Box, makeStyles } from '@material-ui/core';
+import React,  { useEffect } from 'react';
+
+import { Box, styled } from '@mui/material';
+
 import NavBar from './Home/NarBar';
 import Banner from './Home/Banner';
 import MidSlide from './Home/MidSlide';
 import MidSection from './Home/MidSection';
 import Slide from './Home/Slide';
-import React,  { useEffect } from 'react';
+
 import { useSelector, useDispatch } from 'react-redux'; // hooks
 import { getProducts as listProducts } from '../redux/actions/productActions';
 
-const useStyle = makeStyles({
-    component: {
-        padding: 10,
-        background: '#F2F2F2'
-    }
-})
+const Component = styled(Box)`
+    padding: 20px 10px;
+    background: #F2F2F2;
+`;
 
 const Home = () => {
-    const classes = useStyle();
-
     const getProducts = useSelector(state => state.getProducts);
     const { products, error } = getProducts;
 
@@ -30,7 +29,7 @@ const Home = () => {
     return (
         <>
             <NavBar />
-            <Box className={classes.component}>
+            <Component>
                 <Banner />
                 <MidSlide products={products} />
                 <MidSection />
@@ -58,7 +57,7 @@ const Home = () => {
                     timer={false} 
                     multi={true} 
                 />
-            </Box>
+            </Component>
         </>
     )
 }
