@@ -1,49 +1,39 @@
-import { useState, useContext, useReducer, useEffect } from 'react';
-import { Button, Box, makeStyles } from '@material-ui/core';
-import { ShoppingCart as Cart, FlashOn as Flash } from '@material-ui/icons';
-import clsx from 'clsx';
-import { useHistory } from 'react-router-dom';
-import { LoginContext } from '../../context/ContextProvider';
+import { useState } from 'react';
+
+import { Button, Box, styled } from '@mui/material';
+import { ShoppingCart as Cart, FlashOn as Flash } from '@mui/icons-material';
+
+import { useNavigate } from 'react-router-dom';
 import { payUsingPaytm } from '../../service/api';
 import { post } from '../../utils/paytm';
-// import { initialState, reducer } from '../../reducers/reducer';
-import { addToCart } from '../../redux/actions/cartActions';
-import { useSelector, useDispatch } from 'react-redux';
 
-const useStyle = makeStyles(theme => ({
-    leftContainer: {
-        minWidth: '40%',
-        // textAlign: 'center',
-        padding: '40px 0 0 80px',
-        [theme.breakpoints.down('md')]: {
-            padding: '20px 40px'
-        }
-    },
-    productImage: {
-        padding: '15px 20px',
-        border: '1px solid #f0f0f0',
-        width: '95%'
-    },
-    button: {
-        width: '46%',
-        borderRadius: 2,
-        height: 50
-    },
-    addToCart: {
-        background: '#ff9f00',
-        color: '#FFF'
-    },
-    buyNow:{
-        background: '#fb641b',
-        color: '#FFF'
+import { addToCart } from '../../redux/actions/cartActions';
+import { useDispatch } from 'react-redux';
+
+const LeftContainer = styled(Box)(({ theme }) => ({
+    minWidth: '40%',
+    padding: '40px 0 0 80px',
+    [theme.breakpoints.down('md')]: {
+        padding: '20px 40px'
     }
-}));
+}))
+
+const Image = styled('img')({
+    padding: '15px 20px',
+    border: '1px solid #f0f0f0',
+    width: '95%'
+});
+
+const StyledButton = styled(Button)`
+    width: 46%;
+    border-radius: 2px;
+    height: 50px;
+    color: #FFF;
+`;
 
 const ActionItem = ({ product }) => {
-    const classes = useStyle();
-    const history = useHistory();
-    const { account } = useContext(LoginContext);
-    const { id, price, detailUrl, title } = product;
+    const navigate = useNavigate();
+    const { id } = product;
         
     const [quantity, setQuantity] = useState(1);
     const dispatch = useDispatch();
@@ -59,15 +49,15 @@ const ActionItem = ({ product }) => {
 
     const addItemToCart = () => {
         dispatch(addToCart(id, quantity));
-        history.push('/cart');
+        navigate('/cart');
     }
 
     return (
-        <Box className={classes.leftContainer}>
-            <img src={product.detailUrl} className={classes.productImage} /><br />
-            <Button onClick={() => addItemToCart()} className={clsx(classes.button, classes.addToCart)} style={{marginRight: 10}} variant="contained"><Cart />Add to Cart</Button>
-            <Button onClick={() => buyNow()} className={clsx(classes.button, classes.buyNow)} variant="contained"><Flash /> Buy Now</Button>
-        </Box>
+        <LeftContainer>
+            <Image src={product.detailUrl} /><br />
+            <StyledButton onClick={() => addItemToCart()} style={{marginRight: 10, background: '#ff9f00'}} variant="contained"><Cart />Add to Cart</StyledButton>
+            <StyledButton onClick={() => buyNow()} style={{background: '#fb641b'}} variant="contained"><Flash /> Buy Now</StyledButton>
+        </LeftContainer>
     )
 }
 
