@@ -1,43 +1,34 @@
 import { useState, useEffect } from 'react';
-import { Box, Typography, makeStyles, CircularProgress, Button, Grid } from '@material-ui/core';
+
+import { styled, Box, Typography, Grid } from '@mui/material';
+
 import ProductDetail from './ProductDetail';
 import ActionItem from './ActionItem';
 import { useParams } from 'react-router-dom';
-import clsx from 'clsx';
 import { getProductById } from '../../service/api';
 import { useDispatch, useSelector } from 'react-redux';
 
 import { getProductDetails } from '../../redux/actions/productActions';
 
-const useStyles = makeStyles(theme => ({
-    component: {
-        marginTop: 55,
-        background: '#F2F2F2'
-    },
-    container: {
-        background: '#FFFFFF',
-        // margin: '0 80px',
-        display: 'flex',
-        [theme.breakpoints.down('md')]: {
-            margin: 0
-        }
-    },
-    rightContainer: {
-        marginTop: 50,
-        '& > *': {
-            marginTop: 10
-        }
-    },
-    price: {
-        fontSize: 28
-    },
-    smallText: {
-        fontSize: 14,
-    },
-    greyTextColor: {
-        color: '#878787'
+const Component = styled(Box)`
+    margin-top: 55px;
+    background: #F2F2F2;
+`;
+
+const Container = styled(Grid)(({ theme }) => ({
+    background: '#FFFFFF',
+    display: 'flex',
+    [theme.breakpoints.down('md')]: {
+        margin: 0
     }
-}));
+}))
+
+const RightContainer = styled(Grid)`
+    margin-top: 50px;
+    & > p {
+        margin-top: 10px;
+    }
+`;
 
 const data = { 
     id: '',
@@ -58,7 +49,6 @@ const data = {
 };
 
 const DetailView = ({ history, match }) => {
-    const classes = useStyles();
     const fassured = 'https://static-assets-web.flixcart.com/www/linchpin/fk-cp-zion/img/fa_62673a.png'
     // const [ product, setProduct ] = useState(data);
     // const [ loading, setLoading ] = useState(false);
@@ -89,29 +79,29 @@ const DetailView = ({ history, match }) => {
     // }
 
     return (
-        <Box className={classes.component}>
+        <Component>
             <Box></Box>
             { product && Object.keys(product).length &&
-                <Grid container className={classes.container}> 
+                <Container container> 
                     <Grid item lg={4} md={4} sm={8} xs={12}>
                         <ActionItem product={product} />
                     </Grid>
-                    <Grid item lg={8} md={8} sm={8} xs={12} className={classes.rightContainer}>
+                    <RightContainer item lg={8} md={8} sm={8} xs={12}>
                         <Typography>{product.title.longTitle}</Typography>
-                        <Typography className={clsx(classes.greyTextColor, classes.smallText)} style={{marginTop: 5}}>
+                        <Typography style={{marginTop: 5, color: '#878787', fontSize: 14 }}>
                             8 Ratings & 1 Reviews
                             <span><img src={fassured} style={{width: 77, marginLeft: 20}} /></span>
                         </Typography>
                         <Typography>
-                            <span className={classes.price}>₹{product.price.cost}</span>&nbsp;&nbsp;&nbsp; 
-                            <span className={classes.greyTextColor}><strike>₹{product.price.mrp}</strike></span>&nbsp;&nbsp;&nbsp;
-                            <span style={{color: '#388E3C'}}>{product.price.discount} off</span>
+                            <span style={{ fontSize: 28 }}>₹{product.price.cost}</span>&nbsp;&nbsp;&nbsp; 
+                            <span style={{ color: '#878787' }}><strike>₹{product.price.mrp}</strike></span>&nbsp;&nbsp;&nbsp;
+                            <span style={{ color: '#388E3C' }}>{product.price.discount} off</span>
                         </Typography>
                         <ProductDetail product={product} />
-                    </Grid>
-                </Grid>
+                    </RightContainer>
+                </Container>
             }   
-        </Box>
+        </Component>
     )
 }
 
