@@ -1,70 +1,77 @@
 import React, { useState, useEffect } from 'react';
-import { Dialog, DialogContent, TextField, Box, Button, makeStyles, Typography } from '@material-ui/core';
+
+import { Dialog, DialogContent, TextField, Box, Button, Typography, styled } from '@mui/material';
+
 import { authenticateLogin, authenticateSignup } from '../../service/api';
 
-const useStyle = makeStyles({
-    component: {
-        height: '70vh',
-        width: '90vh',
-        maxWidth: 'unset !important'
-    },
-    image: {
-        backgroundImage: `url(${'https://static-assets-web.flixcart.com/www/linchpin/fk-cp-zion/img/login_img_c4a81e.png'})`,
-        background: '#2874f0',
-        backgroundPosition: 'center 85%',
-        backgroundRepeat: 'no-repeat',
-        height: '70vh',
-        width: '40%',
-        padding: '45px 35px',
-        '& > *': {
-            color: '#FFFFFF',
-            fontWeight: 600
-        }
-    },
-    login: {
-        padding: '25px 35px',
-        display: 'flex',
-        flex: 1,
-        flexDirection: 'column',
-        '& > *': {
-            marginTop: 20
-        }
-    },
-    loginbtn: {
-        textTransform: 'none',
-        background: '#FB641B',
-        color: '#fff',
-        height: 48,
-        borderRadius: 2
-    },
-    requestbtn: {
-        textTransform: 'none',
-        background: '#fff',
-        color: '#2874f0',
-        height: 48,
-        borderRadius: 2,
-        boxShadow: '0 2px 4px 0 rgb(0 0 0 / 20%)'
-    },
-    text: {
-        color: '#878787',
-        fontSize: 12
-    },
-    createText: {
-        margin: 'auto 0 5px 0',
-        textAlign: 'center',
-        color: '#2874f0',
-        fontWeight: 600,
-        fontSize: 14,
-        cursor: 'pointer'
-    },
-    error: {
-        fontSize: 10,
-        color: '#ff6161',
-        lineHeight: 0,
-        marginTop: 10,
-        fontWeight: 600
+const Component = styled(DialogContent)`
+    height: 70vh;
+    width: 90vh;
+    padding: 0;
+    padding-top: 0;
+`;
+
+const LoginButton = styled(Button)`
+    text-transform: none;
+    background: #FB641B;
+    color: #fff;
+    height: 48px;
+    border-radius: 2px;
+`;
+
+const RequestOTP = styled(Button)`
+    text-transform: none;
+    background: #fff;
+    color: #2874f0;
+    height: 48px;
+    border-radius: 2px;
+    box-shadow: 0 2px 4px 0 rgb(0 0 0 / 20%);
+`;
+
+const Text = styled(Typography)`
+    color: #878787;
+    font-size: 12px;
+`;
+
+const CreateAccount = styled(Typography)`
+    margin: auto 0 5px 0;
+    text-align: center;
+    color: #2874f0;
+    font-weight: 600;
+    font-size: 14px;
+    cursor: pointer
+`
+
+const Wrapper = styled(Box)`
+    padding: 25px 35px;
+    display: flex;
+    flex: 1;
+    overflow: auto;
+    flex-direction: column;
+    & > div, & > button, & > p {
+        margin-top: 20px;
     }
-})
+`;
+
+const Error = styled(Typography)`
+    font-size: 10px;
+    color: #ff6161;
+    line-height: 0;
+    margin-top: 10px;
+    font-weight: 600;
+`
+// height: 70vh;
+    
+const Image = styled(Box)`
+    background: #2874f0 url(https://static-assets-web.flixcart.com/www/linchpin/fk-cp-zion/img/login_img_c4a81e.png) center 85% no-repeat;
+    width: 40%;
+    height: 100%;
+    padding: 45px 35px;
+    & > p, & > h5 {
+        color: #FFFFFF;
+        font-weight: 600
+    }
+`;
 
 const loginInitialValues = {
     username: '',
@@ -94,7 +101,6 @@ const accountInitialValues = {
 }
 
 const LoginDialog = ({ open, setOpen, setAccount }) => {
-    const classes = useStyle();
     const [ login, setLogin ] = useState(loginInitialValues);
     const [ signup, setSignup ] = useState(signupInitialValues);
     const [ error, showError] = useState(false);
@@ -140,37 +146,37 @@ const LoginDialog = ({ open, setOpen, setAccount }) => {
     }
 
     return (
-        <Dialog open={open} onClose={handleClose}>
-            <DialogContent className={classes.component} >
-                <Box style={{display: 'flex'}}>
-                    <Box className={classes.image}>
+        <Dialog open={open} onClose={handleClose} PaperProps={{ sx: { maxWidth: 'unset' } }}>
+            <Component>
+                <Box style={{display: 'flex', height: '100%'}}>
+                    <Image>
                         <Typography variant="h5">{account.heading}</Typography>
                         <Typography style={{marginTop: 20}}>{account.subHeading}</Typography>
-                    </Box>
+                    </Image>
                     {
                         account.view === 'login' ? 
-                        <Box className={classes.login}>
-                            <TextField onChange={(e) => onValueChange(e)} name='username' label='Enter Email/Mobile number' />
-                            { error && <Typography className={classes.error}>Please enter valid Email ID/Mobile number</Typography> }
-                            <TextField onChange={(e) => onValueChange(e)} name='password' label='Enter Password' />
-                            <Typography className={classes.text}>By continuing, you agree to Flipkart's Terms of Use and Privacy Policy.</Typography>
-                            <Button className={classes.loginbtn} onClick={() => loginUser()} >Login</Button>
-                            <Typography className={classes.text} style={{textAlign:'center'}}>OR</Typography>
-                            <Button className={classes.requestbtn}>Request OTP</Button>
-                            <Typography className={classes.createText} onClick={() => toggleSignup()}>New to Flipkart? Create an account</Typography>
-                        </Box> : 
-                        <Box className={classes.login}>
-                            <TextField onChange={(e) => onInputChange(e)} name='firstname' label='Enter Firstname' />
-                            <TextField onChange={(e) => onInputChange(e)} name='lastname' label='Enter Lastname' />
-                            <TextField onChange={(e) => onInputChange(e)} name='username' label='Enter Username' />
-                            <TextField onChange={(e) => onInputChange(e)} name='email' label='Enter Email' />
-                            <TextField onChange={(e) => onInputChange(e)} name='password' label='Enter Password' />
-                            <TextField onChange={(e) => onInputChange(e)} name='phone' label='Enter Phone' />
-                            <Button className={classes.loginbtn} onClick={() => signupUser()} >Continue</Button>
-                        </Box>
+                        <Wrapper>
+                            <TextField variant="standard" onChange={(e) => onValueChange(e)} name='username' label='Enter Email/Mobile number' />
+                            { error && <Error>Please enter valid Email ID/Mobile number</Error> }
+                            <TextField variant="standard" onChange={(e) => onValueChange(e)} name='password' label='Enter Password' />
+                            <Text>By continuing, you agree to Flipkart's Terms of Use and Privacy Policy.</Text>
+                            <LoginButton onClick={() => loginUser()} >Login</LoginButton>
+                            <Text style={{textAlign:'center'}}>OR</Text>
+                            <RequestOTP>Request OTP</RequestOTP>
+                            <CreateAccount onClick={() => toggleSignup()}>New to Flipkart? Create an account</CreateAccount>
+                        </Wrapper> : 
+                        <Wrapper>
+                            <TextField variant="standard" onChange={(e) => onInputChange(e)} name='firstname' label='Enter Firstname' />
+                            <TextField variant="standard" onChange={(e) => onInputChange(e)} name='lastname' label='Enter Lastname' />
+                            <TextField variant="standard" onChange={(e) => onInputChange(e)} name='username' label='Enter Username' />
+                            <TextField variant="standard" onChange={(e) => onInputChange(e)} name='email' label='Enter Email' />
+                            <TextField variant="standard" onChange={(e) => onInputChange(e)} name='password' label='Enter Password' />
+                            <TextField variant="standard" onChange={(e) => onInputChange(e)} name='phone' label='Enter Phone' />
+                            <LoginButton onClick={() => signupUser()} >Continue</LoginButton>
+                        </Wrapper>
                     }
                 </Box>
-            </DialogContent>
+            </Component>
         </Dialog>
     )
 }
