@@ -1,6 +1,8 @@
-import { Box, makeStyles, Typography, Button, Grid } from '@material-ui/core';
-import CartItem from './CartItem';
 import { useEffect } from 'react';
+
+import { Box, Typography, Button, Grid, styled } from '@mui/material';
+
+import CartItem from './CartItem';
 import { useSelector, useDispatch } from 'react-redux';
 import { addToCart, removeFromCart } from '../../redux/actions/cartActions';
 import TotalView from './TotalView';
@@ -8,46 +10,44 @@ import EmptyCart from './EmptyCart';
 import { post } from '../../utils/paytm';
 import { payUsingPaytm } from '../../service/api';
 
-const useStyle = makeStyles(theme => ({
-    component: {
-        // marginTop: 55,
-        padding: '30px 135px',
-        display: 'flex',
-        [theme.breakpoints.down('sm')]: {
-            padding: '15px 0'
-        }
-    },
-    leftComponent: {
-        // width: '67%',
-        paddingRight: 15,
-        [theme.breakpoints.down('sm')]: {
-            marginBottom: 15
-        }
-    },
-    header: {
-        padding: '15px 24px',
-        background: '#fff'
-    },
-    bottom: {
-        padding: '16px 22px',
-        background: '#fff',
-        boxShadow: '0 -2px 10px 0 rgb(0 0 0 / 10%)',
-        borderTop: '1px solid #f0f0f0'
-    },
-    placeOrder: {
-        display: 'flex',
-        marginLeft: 'auto',
-        background: '#fb641b',
-        color: '#fff',
-        borderRadius: 2,
-        width: 250,
-        height: 51
+const Component = styled(Grid)(({ theme }) => ({
+    padding: '30px 135px',
+    display: 'flex',
+    [theme.breakpoints.down('sm')]: {
+        padding: '15px 0'
     }
 }));
 
-const Cart = ({ match, history }) => {
-    const classes = useStyle();
+const LeftComponent = styled(Grid)(({ theme }) => ({
+    paddingRight: 15,
+    [theme.breakpoints.down('sm')]: {
+        marginBottom: 15
+    }
+}));
 
+const Header = styled(Box)`
+    padding: 15px 24px;
+    background: #fff;
+`;
+
+const BottomWrapper = styled(Box)`
+    padding: 16px 22px;
+    background: #fff;
+    box-shadow: 0 -2px 10px 0 rgb(0 0 0 / 10%);
+    border-top: 1px solid #f0f0f0;
+`;
+
+const StyledButton = styled(Button)`
+    display: flex;
+    margin-left: auto;
+    background: #fb641b;
+    color: #fff;
+    border-radius: 2px;
+    width: 250px;
+    height: 51px;
+`;
+
+const Cart = ({ match, history }) => {
     const cartDetails = useSelector(state => state.cart);
     const { cartItems } = cartDetails;
 
@@ -75,23 +75,23 @@ const Cart = ({ match, history }) => {
     return (
         <>
         { cartItems.length ? 
-            <Grid container className={classes.component}>
-                <Grid item lg={9} md={9} sm={12} xs={12} className={classes.leftComponent}>
-                    <Box className={classes.header}>
+            <Component container>
+                <LeftComponent item lg={9} md={9} sm={12} xs={12}>
+                    <Header>
                         <Typography style={{fontWeight: 600, fontSize: 18}}>My Cart ({cartItems?.length})</Typography>
-                    </Box>
+                    </Header>
                         {   cartItems.map(item => (
                                 <CartItem item={item} removeItemFromCart={removeItemFromCart}/>
                             ))
                         }
-                    <Box className={classes.bottom}>
-                        <Button onClick={() => buyNow()} variant="contained" className={classes.placeOrder}>Place Order</Button>
-                    </Box>
-                </Grid>
+                    <BottomWrapper>
+                        <StyledButton onClick={() => buyNow()} variant="contained">Place Order</StyledButton>
+                    </BottomWrapper>
+                </LeftComponent>
                 <Grid item lg={3} md={3} sm={12} xs={12}>
                     <TotalView cartItems={cartItems} />
                 </Grid>
-            </Grid> : <EmptyCart />
+            </Component> : <EmptyCart />
         }
         </>
 
