@@ -1,39 +1,49 @@
 import { useState, useEffect } from 'react';
-import { Box, makeStyles, Typography } from '@material-ui/core';
-import clsx from 'clsx';
 
-const useStyle = makeStyles({
-    component: {
-        // width: '30%'
-    },
-    header: {
-        padding: '15px 24px',
-        background: '#fff'
-    },
-    greyTextColor: {
-        color: '#878787'
-    },
-    container: {
-        '& > *': {
-            marginBottom: 20,
-            fontSize: 14
-        }
-    },
-    price: {
-        float: 'right'
-    },
-    totalAmount: {
-        fontSize: 18,
-        fontWeight: 600,
-        borderTop: '1px dashed #e0e0e0',
-        padding: '20px 0',
-        borderBottom: '1px dashed #e0e0e0'
+import { Box, Typography, styled } from '@mui/material';
+
+const Header = styled(Box)`
+    padding: 15px 24px;
+    background: #fff;
+    borderBottom: 1px solid #f0f0f0;
+`;
+
+const Heading = styled(Typography)`
+    color: #878787;
+`;
+
+const Container = styled(Box)`
+    padding: 15px 24px;
+    background: #fff;
+    & > p {
+        margin-bottom: 20px;
+        font-size: 14px;
     }
-})
+`;
+
+const Price = styled(Typography)`
+    float: right;
+`;
+
+const TotalAmount = styled(Typography)`
+    font-size: 18px;
+    font-weight: 600;
+    border-top: 1px dashed #e0e0e0;
+    padding: 20px 0;
+    border-bottom: 1px dashed #e0e0e0;
+`;
+
+const Discount = styled(Typography)`
+    font-size: 16px; 
+    color: green;
+`
+
+// component: {
+//     // width: '30%'
+// },
 
 
 const TotalView = ({ cartItems }) => {
-    const classes = useStyle();
     const [price, setPrice] = useState(0);
     const [discount, setDiscount] = useState(0)
 
@@ -53,17 +63,25 @@ const TotalView = ({ cartItems }) => {
     }
 
     return (
-        <Box className={classes.component}>
-            <Box className={classes.header} style={{borderBottom: '1px solid #f0f0f0'}}>
-                <Typography className={classes.greyTextColor}>PRICE DETAILS</Typography>
-            </Box>
-            <Box className={clsx(classes.header, classes.container)}>
-                <Typography>Price ({cartItems?.length} item)<span className={classes.price}>₹{price}</span></Typography>
-                <Typography>Discount<span className={classes.price}>-₹{discount}</span></Typography>
-                <Typography>Delivery Charges<span className={classes.price}>₹40</span></Typography>
-                <Typography className={classes.totalAmount}>Total Amount<span className={classes.price}>₹{price - discount + 40}</span></Typography>
-                <Typography style={{fontSize: 16, color: 'green'}}>You will save ₹{discount - 40} on this order</Typography>
-            </Box>
+        <Box>  {/* className={classes.component}> */}
+            <Header>
+                <Heading>PRICE DETAILS</Heading>
+            </Header>
+            <Container>
+                <Typography>Price ({cartItems?.length} item)
+                    <Price component="span">₹{price}</Price>
+                </Typography>
+                <Typography>Discount
+                    <Price component="span">-₹{discount}</Price>
+                </Typography>
+                <Typography>Delivery Charges
+                    <Price component="span">₹40</Price>
+                </Typography>
+                <TotalAmount>Total Amount
+                    <Price>₹{price - discount + 40}</Price>
+                </TotalAmount>
+                <Discount>You will save ₹{discount - 40} on this order</Discount>
+            </Container>
         </Box>
     )
 }
