@@ -1,4 +1,4 @@
-import { styled, Grid } from '@mui/material';
+import {  Grid, styled } from '@mui/material';
 
 const ImageURL = [
     'https://rukminim1.flixcart.com/flap/960/960/image/2f30db9425df5cec.jpg?q=50',
