@@ -1,69 +1,71 @@
-import { Box, Typography, makeStyles, Table, TableBody, TableRow, TableCell } from '@material-ui/core';
-import { LocalOffer as Badge } from '@material-ui/icons';
 
-const useStyle = makeStyles({
-    smallText: {
-        fontSize: 14,
-        verticalAlign: 'baseline',
-        '& > *' :{
-            fontSize: 14,
-            marginTop: 10
-        }
-    },
-    greyTextColor: {
-        color: '#878787'
-    },
-    badge: {
-        marginRight: 10,
-        color: '#00CC00',
-        fontSize: 15
-    },
-    wrapper: {
-        display: 'flex'
+import { Box, Typography, Table, TableBody, TableRow, TableCell, styled } from '@mui/material';
+import { LocalOffer as Badge } from '@mui/icons-material';
+
+const SmallText = styled(Box)`
+    font-size: 14px;
+    vertical-align: baseline;
+    & > p {
+        font-size: 14px;
+        margin-top: 10px;
     }
-});
+`
+
+const ColumnText = styled(TableRow)`
+    font-size: 14px;
+    vertical-align: baseline;
+    & > td {
+        font-size: 14px;
+        margin-top: 10px;
+    }
+`
+
+const StyledBadge = styled(Badge)`
+    margin-right: 10px;
+    color: #00CC00;
+    font-size: 15px;
+`;
 
 const ProductDetail = ({ product }) => {
-    const classes = useStyle();
     const adURL = 'https://rukminim1.flixcart.com/lockin/774/185/images/CCO__PP_2019-07-14.png?q=50';
     const date = new Date(new Date().getTime()+(5*24*60*60*1000));
     
     return (
         <>
             <Typography>Available offers</Typography>
-            <Box className={classes.smallText}>
-                <Typography><Badge className={classes.badge} />Bank Offer 5% Unlimited Cashback on Flipkart Axis Bank Credit Card</Typography>
-                <Typography><Badge className={classes.badge} />Bank Offer 10% Off on Bank of Baroda Mastercard debit card first time transaction, Terms and Condition apply</Typography>
-                <Typography><Badge className={classes.badge} />Purchase this Furniture or Appliance and Get Extra ₹500 Off on Select ACs</Typography>
-                <Typography><Badge className={classes.badge} />Partner OfferExtra 10% off upto ₹500 on next furniture purchase</Typography>
-            </Box>
+            <SmallText>
+                <Typography><StyledBadge />Bank Offer 5% Unlimited Cashback on Flipkart Axis Bank Credit Card</Typography>
+                <Typography><StyledBadge />Bank Offer 10% Off on Bank of Baroda Mastercard debit card first time transaction, Terms and Condition apply</Typography>
+                <Typography><StyledBadge />Purchase this Furniture or Appliance and Get Extra ₹500 Off on Select ACs</Typography>
+                <Typography><StyledBadge />Partner OfferExtra 10% off upto ₹500 on next furniture purchase</Typography>
+            </SmallText>
             <Table>
                 <TableBody>
-                    <TableRow className={classes.smallText}>
-                        <TableCell className={classes.greyTextColor}>Delivery</TableCell>
+                    <ColumnText>
+                        <TableCell style={{ color: '#878787' }}>Delivery</TableCell>
                         <TableCell style={{ fontWeight: 600 }}>Delivery by {date.toDateString()} | ₹40</TableCell>
-                    </TableRow>
-                    <TableRow className={classes.smallText}>
-                        <TableCell className={classes.greyTextColor}>Warranty</TableCell>
+                    </ColumnText>
+                    <ColumnText>
+                        <TableCell style={{ color: '#878787' }}>Warranty</TableCell>
                         <TableCell>No Warranty</TableCell>
-                    </TableRow>
-                    <TableRow className={classes.smallText}>
-                        <TableCell className={classes.greyTextColor}>Seller</TableCell>
-                        <TableCell className={classes.smallText}>
+                    </ColumnText>
+                    <ColumnText>
+                        <TableCell style={{ color: '#878787' }}>Seller</TableCell>
+                        <TableCell>
                             <span style={{ color: '#2874f0' }}>SuperComNet</span>
                             <Typography>GST invoice available</Typography>
                             <Typography>View more sellers starting from ₹329</Typography>
                         </TableCell>
-                    </TableRow>
+                    </ColumnText>
                     <TableRow>
                         <TableCell colSpan={2}>
                             <img src={adURL} style={{ width: 390 }} />
                         </TableCell>
                     </TableRow>
-                    <TableRow className={classes.smallText}>
-                        <TableCell className={classes.greyTextColor}>Description</TableCell>
+                    <ColumnText>
+                        <TableCell style={{ color: '#878787' }}>Description</TableCell>
                         <TableCell>{product.description}</TableCell>
-                    </TableRow>
+                    </ColumnText>
                 </TableBody>
             </Table>
         </>
